ARM: dts: imx6q-cm-fx6: add sdio wifi/bt
The cm-fx6 has an on-board AW-NH387 Wifi/BT chip that is connected to
the usdhc1 mmc controller. The chip requires an external power supply
as well as a resetline. Enable the controller and add the required
power regulator as well as a powersequence for the reset gpio such
that the chip can be discovered and used.